config: use SimpleKCM as base class
Widget configuration pages in Plasma 6 are required to be KCM components
(see the porting guide). Not doing this may cause issues, and the Plasma
configuration infrastructure (unlike in Plasma 5) no longer adds its own
ScrollView around the pages, so if the windoe is small (or the content
large enough) some options will be invisible without any indication that
something is missing.

Instead, use SimpleKCM as the base.